nlp_transformers_embeddings/requirements.txt
anthropic
openai
python-dotenv
pydantic
numpy
pandas
matplotlib
seaborn
scikit-learn
ipython
jupyter
nbformat
nbclient
pytest
